Reorganize badge URL helper functions; affects [website]
Ref #2698
