+++ core/badge-urls/make-badge-url.js
