+++ core/badge-urls/make-badge-url.spec.js
